Add files via upload
This file plots TL according to our study, and you can compare it with what ASCE 7 uses. It uses two shape files to plot the states and overall border of the conterminous US (CONUS), and it also uses a .csv file of modal magnitude deaggregated from spectral periods of 2s and 4s.